+++ src/main/com/study/repository/RequestDomainRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by roy on 2017/8/2.
  */
-public interface RequestDomainRepository extends JpaRepository<RequestDomain,Integer> {
+public interface RequestDomainRepository extends JpaRepository<RequestDomain,Long> {
 
 }
